+++ package.json
@@ -8,6 +8,7 @@
     "material-ui": "^1.0.0-beta.45",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.4",
     "typeface-roboto": "0.0.54"
   },
